fix: ix_index u8 → u16
Spec §9 Domain event taxonomy types SolverInteraction.ix_index as u16.
Bump both the event field (events.rs) and the upstream decoder view
(ResolvedInstruction.ix_index) to u16 for consistent typing through the
pipeline; the single downcast from the grpc u32 lands at the future
construction site.

+++ crates/solana-indexer/src/types/events.rs
@@ -100,7 +100,7 @@ pub(crate) enum SettlementEvent {
         /// Transaction signature.
         tx_signature: Signature,
         /// Index of the instruction within the transaction.
-        ix_index: u8,
+        ix_index: u16,
     },
 }
